requirements.txt
DateTime
uuid
pysqlite3
pymorphy2
pytrovich
comtypes
python-pptx
openpyxl